% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chwd2curr.R
\name{chwd2curr}
\alias{chwd2curr}
\title{CWD to current file's dir}
\usage{
chwd2curr(fnc, verbose = TRUE, fatal = FALSE)
}
\arguments{
\item{fnc}{A function that takes no argument and returns nothing (\code{function(){}})}

\item{verbose}{Whether to print debug messages}

\item{fatal}{Whether to call stop() if we can't get the directory of the currently running script}
}
\value{
None
}
\description{
Change the current working directory 
to the directory in which the script we are running resides.
}
\examples{
chwd2curr(function(){})

}
