% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2dt.R
\name{df2dt}
\alias{df2dt}
\title{Convert a dataframe to DT::datatable for visualization}
\usage{
df2dt(
  df,
  to_truncate = "auto",
  truncate_max = 20,
  truncate_max_title = 200,
  types_in_colnames = TRUE,
  filter = "top",
  to_round = "auto",
  to_round_digits = 2,
  fixed_header = TRUE,
  page_length = 20
)
}
\arguments{
\item{df}{A dataframe}

\item{to_truncate}{Text columns too long that needs to be truncated}

\item{truncate_max}{An integer with maximum length of a string before it is truncated}

\item{truncate_max_title}{An integer representing the maximum length of the substring in tooltip}

\item{types_in_colnames}{A boolean equal to TRUE if datatypes should be appended to column names}

\item{filter}{Position of the filter widgets}

\item{to_round}{A character vector with numeric colum names that have to be rounded}

\item{to_round_digits}{Maximum number of digits for rounded columns}

\item{fixed_header}{Whether the header row should be fixed when scrolling}

\item{page_length}{Number of rows per page}
}
\value{
a DT::datatable
}
\description{
Convert a dataframe to DT::datatable for visualization
}
