% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtss.R
\name{dtss}
\alias{dtss}
\title{Open a shiny server for server-side DT::datatable visualization}
\usage{
dtss(dt)
}
\arguments{
\item{dt}{DT::Table}
}
\description{
Open a shiny server for server-side DT::datatable visualization
}
