% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Compute the mode of a vector
see: https://stackoverflow.com/a/8189441/1719931}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the mode
}
\description{
Compute the mode of a vector
see: https://stackoverflow.com/a/8189441/1719931
}
