% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewdf.R
\name{viewdf}
\alias{viewdf}
\title{View dataframe with Javascript Datatables library
with server-side rendering}
\usage{
viewdf(df, ...)
}
\arguments{
\item{df}{A dataframe}
}
\value{
None
}
\description{
View dataframe with Javascript Datatables library
with server-side rendering
}
