% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewdf2.R
\name{viewdf2}
\alias{viewdf2}
\title{View dataframe with Javascript Datatables library
with server-side rendering}
\usage{
viewdf2(df, ...)
}
\arguments{
\item{df}{A dataframe}
}
\value{
None
}
\description{
View dataframe with Javascript Datatables library
with server-side rendering
}
